requirements.txt
requests==2.26.0
pytest==6.2.5
pytest-allure-adaptor==1.7.12
allure-pytest==2.13.3
